V 3.0.62 : Correction - Amelioration Correction : saisie de la valeur du point de consigne possible jusqu'a 99999999 au lieu de 99 pour utilisation Diluteur API 700 -> modification structure BDD Amelioration : gestion des filtres equipement : possibilite d'utiliser touche entree pour valider

+++ CreationBaseMySql.sql
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS `Concentration` (
   `id_Concentration` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `id_systeme_etalon` int(10) unsigned NOT NULL,
   `id_molecule` smallint(5) unsigned NOT NULL,
-  `point_consigne` smallint(6) NOT NULL,
+  `point_consigne` int(10) NOT NULL,
   `conc_reelle` decimal(10,2) NOT NULL,
   `conc_ozone` smallint(5) unsigned NOT NULL DEFAULT '0',
   `archivee` tinyint(1) DEFAULT NULL,
